Share permission lists across core roles
